backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-multipart==0.0.20
pymupdf>=1.25.0
httpx==0.28.1
openai==1.78.1
python-dotenv==1.1.0
aiofiles==24.1.0
pydantic==2.11.4
